chore: update .gitignore and refine README.md
Since this repo is picking up some traffic, I've thought might as well
update README.md with (pretty!) badges and rewrite some parts of it to
be less formal (because to be honest I was just practicing formal
markdown documentation writing, and most of the stuff I wrote was just
kind of verbose and no one wants to read a wall of text, README's
should be playful :3).

The .gitignore update is just a bonus. It's more complete now, but also
unnecessarily so. I've added ignores that I don't need, but maybe
someone else will find them useful.